Add passportjs authentication support.
Initial commit providing passportjs (http://www.passportjs.org)
authentication middleware.  Provides a single option 'local'
that can be used for debugging.  It stores users locally in
the database with a password (using bcrypt for hash and salt).

Added passport-local and bcrypt as development dependencies
since the local strategy is only intended to be used in
dev/test mode.