Project based updates and refinements to layouts and view rendering.
This checkin contains some udpates to model saving logic, and seperation of conc
erns therein.  It also contains the addition of Twitter Bootstrap v3.0, and the
implementation of a header in said bootstrap fashion onto the layout view (serve
r side).  This allows us to have a constant nav bar that we can use and also it
serves the purpose of giving us a 'container' for our single page app to live in
side of (aka below this nav bar).  This checkin also introduces a few new files to deal with the show view as well as preliminary changes needed for a major overhaul of how we are doing our view rendering on the client side.  We are going to be switching to psudo-master views within the home view and a move away from URL's for the time being, so there are some changes related to that.

 enter the commit message for your changes. Lines starting