Updates on the URL's.
I changed the URL's hash to change upon link click going from the list view to the show view, and it updates itself using the ID of the item.  So it goes from /projects to /projects/:id.  So that follows convention now.  And to get back we re-instantiate the collection and filter through them.