Add OAuth2 support for login/auth
OAuth2 authentication has been added using PassportJS.
The 'oauth2' endpoint and passport strategy is intended
to be used for development and testing, but not production.
It provides an example of what an OAuth provider might do,
including storing keys and creating users.

The OAuth2 setup is intended to talk to a forked version
of oauth2orize available at:
https://github.com/Innovation-Toolkit/oauth2orize-midas-testserver