Added the ability for a user to create a task, once inside of a Project Show View.
There are also various refactorings here, as well as some minor changes to the way we are rendering things.  There still needs to be a larger overhaul of our instantiation system, but in the mean time I am experimenting with what uses less memory.  For now the pattern Joe and I discussed on the whiteboard seems to be working, but I have a feeling it won't be sustainable.  We will see though.  As for the other changes in here they are rather minor related to project show views and other areas related in order to get Tasks to be created and listed.