Removed URL for Task List from Router.
I removed the ability for a user to navigate to the URL for tasks individually from projects for now.  Typically that route would list all the tasks available in the system, but also in order to have restful routing we need to change a few things so for now we are just going to remove that route.