Updated the homepage route.
Using an empty string as a route in backbone, unfortunately, will turn any server side views into the home template view by over-writing the rendering process.  But So I changed it to a slash (/) which is what the default is, though with using a slash there is also some issue with -not- loading, in this case.  So there are issues with both, but the one that doesn't clobber what we are working on is the better.  I have it down to research this more.