Add file upload/find/download support
New FileController allows files to be uploaded and associated
with a user.

Lots of new policies to ensure that arbitrary files cannot
be downloaded. Can read files and file metdata if 'isPrivate' is
not true. Can only post, delete, and put files
(although PUT is an odd activity and needs to be tested/implemented)
if you are logged in.

Added an admin policy, since admins should be able to do anything.
Expose more endpoints for admin users (such as searching users).

Disabled the main 'index' action for files so that users cannot list all
files. Require a file id for everything except a post request to create.

TEMPORARY:
- Include a test function to put a file in the DB from disk.
- Include a test view (testupload) to illustrate how uploads work.