addUserId policy now inserts userId param
To take advantage of the sails blueprints, we can inject the userId
during the policy processing.  So when /create is called, put the
userId in the body of the request (if specified), else in the
params.

With this change, tasks and comments automatically get the id of the
logged in user.