Associate projects with owners
When calling /find on a project, return the owners of the project
in the `owners` field.

Allow an owner to add another owner via a POST to /projectowner.

Allow an owner to delete an owner via DELETE to /projectowner or
/projectowner/destroy.  Be sure to use the `id` of the index in the
projectowner table (found in project.owners.id when you call find)
in order to perform the delete operation.