Refactoring on Projects and Nested comment rendering.
I did a bunch of refactoring around the project models and rendering process, and unfortunately
I introduced a bug around the -necessary- refactoring.  So now before saving a project we need to do a page refresh.  It will still automatically render it to the page, but we need to do a page refresh in between each.  This is due to the newly added save for photos as well as the save for creation so a POST and PUT are being mixed up.  I will handle this in demo though.