Set up API (backend) test framework.
Tests for the backend are now in test/api.  They are run by mocha and
the npm test (or make test) command.

This is a work in progress, and there's still a few issues
(like sailsjs spawns grunt, and for some reason the grunt process
doesn't die when sailsjs dies)