Some code cleanup, and the beginning of more advanced functionality.
This commit includes some more advanced set up for the way we handle events
that are being delegated from within another controller.  So for instance if we have
/projects/1 -> and we have a show controller respectively.  That controller shouldn't be
responsible for initializing the tasks list controller, as that is more of a project_app
global state check and display method.  So thereby we set up a new event object for sending render
states around the app.  Very similar to $(document).ready(), but in this case
we are checking render states of sub-views within our top level views.  This helps with the
clearing out of views later as well.