New test framework
Migrate to using require('sails') for tests.

The sails.upstart.js script loads a sails instance, and then
keeps it around until all of the tests have run.  In this way,
each of the scripts doesn't need to start their own instance
of sails, which was problematic.

Remove the app.test.js script because it does the same
functionality as the upstart.

Update user.test.js to use the new framework.

Update makefile to call mocha --recursive