Auto-complete search feature
Auto-complete enables queries for a particular topic in different
contexts.  The initial contexts are:
- search : the main search box
- inline : searching in any text box, such as @user
- attachments : searching for documents to attach

This commit has the basic structure for adding sources of search
data and then querying those sources for the desired term.

The endpoint is
/ac/:action?q=:term

For example:
/ac/inline?q=joe

Sources are defined in `api/services/sources`.
The configuration is at `config/settings/sources.ex.js`.

In each config, it specifies a `type`.  That type is then
used to look up the correct source module, and query it for
the search results.  The full config is passed to the
source module to do with it as it pleases.  It returns
the tuple `(err, results)` where a non-null error object
means something went wrong.

Initial support permits searching fields in sails models,
and search mediawiki instances like wikipedia.