Add 'likes' for projects
Add the ability to like individual projects.

Implements a `Like` model that stores the projectId and
userId association.

Implements a `LikeController` that allows certain operations
such as `find` and `count`.  Utility functions are also exposed:

/like/like/:projectId will cause the current logged in user to like
the project specified

/like/unlike/:projectId will cause the current logged in user to
abandon their like.

Added a test suite.

Used 2 new policies: requireId (requires an `:id` parameter to be
specified in the URL) and requireUserId (requires that the
user is logged in)