Updated a function on the Utilities mixin.
I updated the function getCurrentModelFromFormAttributes to actually
behave in the way that the name expects.  Now you can grab any attribute
from any model on the front-end (within a collection), then pass it back
to this function, it will look to see if that attribute exists within
each model, and if the value of the attr matches return back out that
model.