UI framework for tags using select2
This is an initial commit for adding a select2 box to pick tags
for a project.  select2 has been added as a reference to the main
repo via submodule.  require_config.js has been updated to
import select2, which is NOT an AMD module (thus the added shim).

The framework is in place to search for tags, but as commented,
the handler functions in the controller are not yet implement.
Nor are any custom render functions for the select2 results.