Add modal footer; scaffold for tag modal
The tag modal allows an owner/admin to create a new tag.
This commit just adds the scaffold for showing the form,
but does not implement the logic on submit.

Modals could not have a modal-footer part of their templates,
so instead put in a modal-template div and attach the rendered
form to that item so that both modal-body and modal-footer
can be defined.  There's probably a better way to do this,
but this is a quick fix.