Update to Events Scaffolding.
- Events creation and dynamic adding now works.
- Events RSVP and Canceling of RSVP now works.

Note: I had to remove the UUIDV4 from the Event model on the back-end due to not knowing how to work with that yet.  Joe, I will simply uncomment that once we have a quick discussion on how I should use that.  And yes Im aware its probably in a commit message somewhere but I looked can't find it.  No biggy a 2 min convo will fix it.

- Dain