Functional adding of tags and dynamic render
Add a new template for rendering tag contents, and
then allow created or selected tags to be stored on the
backend and rendered in the appropriate tag type.

Add a new 'tag' backbone component to store configuration
for tags.  Since tag names, types, etc are mainly a
end-user customer choice, let them decide which tags
are there by editing the tag.js component in backbone.
The backend doesn't need to know anything about the
type or content of tags, it just provides the machinery
for the front end.

Add new items to require, and add the async module
so that we can do multiple calls and then return
on their completion.