Connect UI to comment topic feature
Make changes to the UI, UI views, and UI controllers to
support the notion of discussion topics.

The "new topic" button drops down an input box for the
topic content.  This is then saved via an AJAX call and
the collection is refreshed.  DM: please fix this to
use the appropriate collection calls if it shouldn't work
this way, which is outside the backbone framework.

The comment-list-wrapper now renders all of the comment
UI, including the header.  Inside of each comment-list,
when there's a topic, it renders the topic and then
the comments for that topic, then a placeholder for
the comment form at the end.

After the view is rendered, the new comment sub view
is rendered for each of the topics and inserted into
its comment-form-wrapper.  The new comment view has
been modified to be passed the topicId (parentId)
as an option, so it can relay that to the backend.