Create utils; update comments to support tasks
Created a new directory called `utils` in services in
preparation for tasks.  Include project.js and task.js
that provide the `authorized` function for checking whether
a given userId is authorized to view a particular project
or task.

Updated CommentController by refactoring the comment query
and processing code into a helper function which can be called
by both findAllByProjectId and findAllByTaskId.