Refactor tags to app (tags/show)
So that tags can be reused in multiple views, refactor the
tags functionality into a separate app called tags/show
and create a backbone view called TagShowView.

This is the first step to reusing the tag functionality
in tasks (and user profiles).  The tag app does not yet have
the hooks to configure it for a particular view (eg project
or task), but it can now be instantiated from any view.