Navigate using Backbone routes
It was driving me batshit crazy that navigation
has been broken for this long.  When you click on "Browse
Projects", it should let you browse the projects.  When you
click on a project, the URL should change.  When you're
viewing a project, and you hit reload in your browser, it
should reload that project.  When you click forward or back,
the browser should go to the previous or next page.

This commit fixes all of this by implementing routes in project_app.js.

Any time we change to a new view, such as project show, task show,
profile, etc, we should be using the router.navigate() function
and not directly rendering the view/controller.

This fixes Issue #4.