Add rigor to cleanup() and render()
For each controller, have the controller call cleanup on
any of the views that it has instantiated.  Add a utility
function to standardize how cleanup gets done across views.

Change rendering.on() to this.model.on().  Without this,
every time a view is generated, it listens to rendering.  Then,
it competes with the other instantiated views to render the
same DOM elements.  The result is chaos.  Only the model
currently being displayed should get the notification
about rendering.  The good news is that it appears
triggering events on this.model are limited to those
objects which are specifically listening to that model
object, and not other models.