Use the /:path instead of #:path
With turning on pushState, paths can get corrupted with
things like /projects/1#projects by using the hash method.

The downside with this approach is that the whole application
reloads because <a href=> tags that have '/' paths get loaded
again from the server (layout.ejs).

The correct (down the road) fix is to have all of the HTML
<body> content rendered within Backbone, and use the
route.navigate() function whenever one of the navigation
bar buttons is pressed (making a `nav` application)

Probably should file a GitHub issue for this.  TBD.