Crop profile images to a square on upload
Profile images as circles look weird if the image itself
is not a square.  This could be solved in two ways:
(1) using <canvas> elements every place a profile
image is shown
(2) cropping the image on the backend when it is uploaded
so that it is square.

I chose to do the latter because it:
- Reduces the amount of data stored
- Is quicker to render (less compute on server and client)

This update REQUIRES two important things:
* Installation of GraphicsMagick
  sudo apt-get install graphicsmagick
* Installation of gm npm package
  npm install