Merge of Sprint6-Comments Branch w/S7.
This marks a time when I finally was able after trying many times to
merge sprint6-comments into sprint7.  I had a corrupted git package of
some sort (I think), but the point is I wasn't actually able to show the
/routes as working appropriately on my local, and then once I fixed that
I wasn't able to see my comment code as working properly on S7.  To fix
this I blew away S7 entirely, re-cloned and got a new fresh copy of the
branch for S7 and then brought over my code that I had changed manually.
This is all that I can see that needs to be brought over for comments.