Added reusable concept for task modal.
The task modal needs to be a wizard-esque step by step process.  And it
needs to be modifiable in the future.  Thus, simplicity rules and a way
to iterate through any current item to the next one, and so forth is the
ideal way to handle this (imo).  The concept means we don't save the
state of what item we are on (in digit form), but we do with a current
flag.  That flag is what allows us to push in error messages, and handle
switching logic.
Added reusable concept for task modal.

The task modal needs to be a wizard-esque step by step process.  And it
needs to be modifiable in the future.  Thus, simplicity rules and a way
to iterate through any current item to the next one, and so forth is the
ideal way to handle this (imo).  The concept means we don't save the
state of what item we are on (in digit form), but we do with a current
flag.  That flag is what allows us to push in error messages, and handle
switching logic.
Added reusable concept for task modal.

The task modal needs to be a wizard-esque step by step process.  And it
needs to be modifiable in the future.  Thus, simplicity rules and a way
to iterate through any current item to the next one, and so forth is the
ideal way to handle this (imo).  The concept means we don't save the
state of what item we are on (in digit form), but we do with a current
flag.  That flag is what allows us to push in error messages, and handle
switching logic.
Added reusable concept for task modal.

The task modal needs to be a wizard-esque step by step process.  And it
needs to be modifiable in the future.  Thus, simplicity rules and a way
to iterate through any current item to the next one, and so forth is the
ideal way to handle this (imo).  The concept means we don't save the
state of what item we are on (in digit form), but we do with a current
flag.  That flag is what allows us to push in error messages, and handle
switching logic.
se enter the commit message for your changes. Lines starting