Make popovers stick for a little while
Popovers should stay up when the user keeps their mouse pointer
in the popover.  This allows us to (in the future) put buttons
and links into the popovers (such as email, go to profile, star,
etc).

Eliminates the need to hide the popover from the controllers too,
which makes it easier to implement; just need to set up the event
to trigger the popover going on.