Fix Issue #7 - load new profile image
When storing a new image on the server, the server
would often cache the image.  Force jQuery to load
the image, which invalidates the cache, and refreshes
the background.

At the same time, fix a bug where the 'edit' id was
being passed to fetch the model, which is clearly an
error.