Make the tags ordered so you can sort
You may want the tags to be rendered in a certain order
in the UI.  Since associative arrays have no order, instead
use a list/array which does have an order.

This allows the UI, when it gets all the tags of a given type,
to sort the list by id and thus get the display order correct.