Comments WORK!
Major overhaul of comments to get them to work.

Centralized a lot of the comment functionality so that
the same code is always being called for things like add.

Use more robust css selectors to find DOM elements
expecting that lots of the middle styling DOM aspects will change.

Use the removeView() function in cleanup to correctly
clean up views.  Keep a list of all of the views instantiated
and remove them when the page needs to cleanup()

restructure the html/css to look kinda (but not really) pretty

remove unused requirements from controllers/views