Show projects in profile
This is a first checking of showing all the projects that you
own in your profile page.  It is designed as an 'activity app'
so it will be reused to show your tasks.

There's a few things I want to do, like configurable columns
passed as parameters to the activity view (for example, tasks
may want to show a deadline).

Also an open question how to format the dates.