Added new paradigm for Ajax Spinner, details below
There is now a base_view class that every RENDER ONLY child view should
extend off.  This base_class will automatically render the view once it
is ready to do so.  In this case, once it has set the ajax spinner and
attached it to the $el of the child class.  The child class then has its
render method called in which it flushes the DOM and adds new content to
its $el.

Definition of render only:
- It only renders, and that's it.
- NOTE: We have been adding more logic than is needed into the views
  themselves, and with this paradigm I hope to slowly get us back to
  using the controller more.  This is just the first step in bringing as
  much generic functionality as we can to the base_view class.

This means we no longer need to do this:

var viewName = new ViewName({
  el: "#someEl",
  model: someModel
  // ...
});
this.viewName.render();

We can simply do this:

var viewName = new ViewName({
  el: "#someEl",
  model: someModel
  // ...
});

And it will automatically know to initialize logic within the base class
with a reference to the view you call new on, and then render the
immediate child view of the base class (current instance initialized).

Based on analysis of our application, there is no situation in which we
initialize a view class but don't render it right there.  Sometimes
there is logic within the view if it is large, and thereby making it NOT
a -render only view-, in which the render method is the second, third,
or fourth method called within said view.  That class for now would not
be a good one to extend the base_view class on, as I mentioned above it
has to be a render only view class.  But eventially I'd like to check
for that and/or pass in a param to the constructor that allows the
parent class to know how and when to render its child view.