Added Production methods to the Gruntfile.
I added a production compile method to the grunt file, and the deps for
it in the package.json, tested, and built our project and stored it in
/assets/js/prod/midas.js.  We can change that directory we house this
in, but for now that seems fine.

To run this command simply open terminal, do an npm install, and
preferably npm install grunt-contrib-requirejs --save-dev, then type
"grunt requirejs" to execute the requirejs grunt command, and you will
see the app compiling.