Updated Tasks Show View Logic
Updated the logic and initialization on the task show view to behave the
same as Profiles (to an extent).  There is now the ability to render the
view state (tags, task specific data, etc) and then edit that as well,
fetching all the data needed right there in the show view.  Eliminating
the need to do it again in the edit view.  Template files have yet to be
checked in, still working out bugs.