Get requirejs to work in production.
This was a pretty hairy ugly undertaking.  Our requirejs
app was really not set up for use in dual-mode development
and production.  I've hacked the app to look more like a
well-defined require module.

I've then modified layout.ejs to check what environment
sails is running in (development or production)
and render to the page the correct <script> html.

When you lift sails, it does NOT generate the production
javascript... instead you need to do:

	grunt requirejs
	sails lift --prod

This will run the server in production mode with a single
javascript file (other than jquery).

There's still more work to be done here -- like figuring
out how to include jQuery in the configuration rather than
hacking it in as a <script> tag, but this seems to run
well locally.  The other thing would be to compile
the production js on sails lift (And also inspect what
sails is really doing during production to minimize
cpu and memory resources hogged by grunt)

Still need to test it on the remote demo server.