Experiment with moving tasks/events
Put tasks/events in the main column, and attachments on
the right.  This layout is not great, but better than
things in the sidebar.

I have an idea for how to make this MUCH much better
but not enough time to implement it before today's
demo.  So checking this in since it is an improvement to
the way it was.