Migrate tasks to a real route
Put tasks at /tasks/:id

Update the backend to handle this case.

Create a router for tasks that handles
routing in backbone for /tasks/:id.

Update model to be able to fetch and notify
when fetch is done.

Update the route in the project show page
(task_list_controller) to navigate to the task.