Remove bower for now
Everything is committed to the project and we're not using it,
so removing it for now.  Should we remove the frontend libraries
from the git repo, and instead decide to use bower, we can bring
it back.  But for now it is woefully out of date and not
being used.