Abstracting Submit Event on Forms inside Modal
I've started moving the submit event (and logic) out of the
form which is a child of the ModalWizard class.  This will now allow and
take any data, target, and modelName that you pass in to the modalWizard
and at the end of the steps available that button will be actionable.