Major design overhaul
The existing design wasn't working for me, so I wanted
to change it to look a bit more modern while still using
splashes of color (but not 27 colors!)

This moves a lot of the controls from the top or left
to a sidebar on the right of the project.  This includes
like, edit, email, calendar, etc.

Task and Event views get a more list-like treatment inline.
They are bounded to a max-height, and become scrollable
div's after that.  Tasks are allowed to be larger than
events.

Buttons are hidden if you are NOT logged in, so you can't try
to add a task or post a comment (that was confusing me, so
it made sense to just remove those DOM elements when the
user is logged out)

CSS is consolidated into a series of general styles for
boxes and padding (box, box-pad-lr, box-pad-t, box-main)
Color treatment still occurs in the app-specific css
(in this case, innovation-theme.css).  So all your colors
can be changed there.

Not everything is hooked up in Backbone, but everything
that worked before should still work (only new stuff needs
implementation)

The task show page did not get the new treatment because
that page seems like it is in a major state of flux and
was not merged with sprint8

Github issues to file:
- Comments need to roll up and expand on click
- Event rsvp button doesn't seem to work
- DK's work on making project edit like profile edit
- Say 'no discussion yet' when no discussion.
- Move fonts into the project instead of linking
  to Google's servers (for off-network deployments)
- Create a default project picture for both
  show and list views
- Icons next to each main section in project?