Search is functional!
Backend now renders search results,
and UI displays either search results or the
project collection depending on how many
search terms are specified.

The search right now is an 'OR' search.
I know that's not what we want -- we want
to 'AND' the terms.  That was a lot more work.
So for this first version, the terms will be
'OR'd together, and probably an option later
will be to choose either 'OR' or 'AND' functionality.

But it works and renders correctly.