Add 'Browse Opportunities'
Based on a discussion with Mollie, use the browse app
to browse through different types of things.  In this case
you can now browse opportunities.  Opportunities can be
filtered by the tags that apply to tasks (although this
has not yet been added to the tag config in tag.js)

The styling for task "cards" is not yet implemented.
This is just a POC and will be filled out shortly.