Clean up RSVP button fns
The jquery element selector was just wacky and totally
dependent on the HTML structure.  Fix that by looking for
the div.event parent, and then the appropriate children
from there.

Correctly render the button, and include a utility function
to update the # of people going immediately.

Add a cancel RSVP hover state to provide more feedback
when you're going.