Consolidate routers
With multiple routers, it is almost impossible to
cleanup views/controllers when navigating between them.
Failure to do so leads to multiple events being triggered
across controller and ultimately faulty logic.

With everything in a single router (we're not there yet,
but that's the goal and the recommended Backbone paradigm),
we can correctly cleanup, reset data, and instantiate the
right controllers.

This fixes a bug with likes buttons triggering all kinds
of chaos on pages you previously visited in your
backbone history.