Add task cards to browse view
This is a REALLY basic view of tasks and their tags.
The framework is in place for this to be a LOT cooler,
but I'm really blocked by the task-new stuff (like setting
locations, sponsoring offices, the need for task tags
to be in tag.js, and the renaming of tag types to be things
like task-people-required).

Also implemented popover clicks, so you can now click on
anything that has a popover and go to that person's profile.

If you're not logged in, you get a blank page.  This will be
fixed once the login system is moved to a backbone app instead
of a server side view.