Update build system
Modify CSS to be compiled for the project for production,
minimizing the number of files transferred and minimizing
the size of the resulting css.

Update grunt to build correct all of the targets.  You
now build the production environment files with:

	grunt build

Modify the sails lift --prod grunt rules to only
copy the necessary files (and hopefully speed up the
process of sails lift in production, especially on EC2)

	sails lift --prod