Attempting to write unit tests for Task section.
@polastre, can you pick up where I left and help me finish this test.
I'm writing a test so that I know there will be a baseline of tasks that
have tags, which I can use to test against.

I can't see a difference between sprint8, and mine.  I just caught it up
again.  I will include a screenshot of what I am seeing in this commit
in the task section.  The live data is working.

JUST AS A NOTE: For this commit/question/test we are only referring to
the madlib on the task show page that -should- be populating the tags
on said page.

I do need your help though to finish up the tests Joe, as I think once
we at least can verify some test data across the dev environments then
we can see what the problem is.  As I am 99% sure it is an issue with
the type of data being returned/attached to a task.