Overhaul task creation and task tags
Refactored all of the task tags to start with `task-*` so
we can tell which tags are task specific.  This affects
the tag.js configuration file, as well as the init build config.

The wizard modal was having issues, mainly because it wasn't
being cleaned up correctly.  Add the right checks, clean up
the subviews, and move the modal out of the
task_collection_view_template.  Put a wrapper for the modal
in the project_item_view_template, and that way the modal
doesn't get inadvertantly destroyed by an update of the task
list.

Fix lots of modal wizard bugs:
- Forward and back now work
- Now detect correctly when at start and end of the wizard
- Clean up the modal wizard and its sub view
- Cancel button works
- Remove page refresh hack
- Submit button works, and offers sub-views the chance to
  change the data model before being submitted
- Add a listener to hide the modal

In task_form_view.js:
- Simplify the logic for tags
- Create the list of tags correctly by either merging lists
  or pushing to the existing list depending on object type

Net:
- Tasks create
- All task tags are added
- Task list view refreshes after task is added
  without page reload
- Modal wizard view destroys itself

TODO:
- Location should be selected as a tag.
- Refactor other task views to use new tag names from tag.js