Updated /task/edit to take a route.
Now there is a route for /task/:id/edit and the route gets initialized
everytime the user goes to click on edit.  The DOM then will update the
button to 'show' alerting the user they can click to go back to the show
view, but if they do they will lose any unsaved changes.  Then the out
of scope click event ('viewEvent' I called it) will trigger the route
for the show view.