Overall task show page
Task show page now looks and feels like the rest of the app.

The tag view now works correctly, and ripped out the previous
tag view that was specific just to tasks.

The comments now work, and you can pass in a target such as
'task' to the comment controller.

The task render() is dependent on the model fetch, which means
the page is blank until the model loads.  There was an order
of operations problem where tags and comments would try to
render before the task show view rendered (due to waiting
for the fetch).  Trigger an event so the controller knows
when it is safe to render the rest of the page.

Added back the topic slide so that topics move up and down.

Fix a number of bugs in the extension of the comment
controller to be reusable for both tasks and comments.

Much more TODO:
- like button
- attachments
- figure out why task show removes #container during cleanup
  (that's real bad)
- sidebar buttons
   - open/close task
   - toggle edit view...
   - volunteer