Fill task edit with existing selections
Correctly fill each of the select2 elements in
the task edit view with the correct values (those
being the values that were selected previously for
this user).  This wasn't working before, but seems
to work now.

Location needs to be fixed as it isn't consistent
with the new task view, and the new task view should
ask for a location from the list.