Task save/update logic
Figure out the "diff" of the tags currently selected
in the edit view from the tags that were previously
selected/stored.  Delete the ones that were
removed and add the ones that were added.
Update model metadata as well.

Redirect back to the task once the edit finishes saving.

With this, task edit view is finally functional.