Move login to a frontend backbone app
Login/authentication was being served by the sails backend
in previous revisions.  This commit moves the login functionality
to be a frontend backbone application, initiated by the navigation
app.

To enable this, needed to pass user credentials in the layout.ejs
from the backend if someone is logged in.  Then use those credentials
to render the page.  The app is initiated if login is clicked,
and then stores the user credentials for other pages to use
in the `window.cache.currentUser` variable.  Added a
`window.cache.userEvents` object to trigger new events
like user:login and user:logout.