Trigger login on unauthorized page
If you reach a page where you're not logged in, or you
decided to log out on a page that you're uniquely able to
view, trigger the login dialog.  Currently implemented
for the profile page.

Trigger `user:request:login` and the login dialog will appear.

You can pass a custom message to the trigger if you'd like to
inform the user of why you're showing the dialog.  Such as:

window.cache.userEvents.trigger('user:request:login',
	'You must be logged in to view profiles.');

Added an alert template so that alerts can be rendered by
a page instead of a blank canvas.