Pass LinkedIn fields; fix callback url bug
Need to extract in a common manner the LinkedIn profile fields
that are useful for setting up a user's account.

There was also a bug where the callback would always be specified
to be localhost.  Instead, use the settings from
config/settings/auth.js which was previously not being used.