Store email (local) and providerinfo (oauth)
If logging in for the first time with a local
account, check if the username is an email address
(in most cases it will be if the front end validates
that the username is an email address).  If it is, then
add it to the list of email addresses for that user.

If adding a service and it has more info about the
person, copy over those fields to the person's profile,
such as title, bio, etc.

No support yet for copying tags (location, company,
skills, topic)